// Front-end Application - Angular/src/app/models/library-item.ts
export class LibraryItem {
    
    title: string;
    datePublished: string;
    publishedDate: any;
    itemType: number;
    borrower: string | null;
    borrowerNo: string | null;
    sector: string | null;
    isbn: string | null;
    noOfPages: number | null;
    authors: string | null;
    languages: string | null;
    subtitles: string | null;
    producer: string | null;
    actors: string | null;
    id: string | null;
    dateBorrowed: string | null;
    lateFee: number | null;
  }